+++ custom-typings/path-is-absolute.d.ts
@@ -1,6 +1,0 @@
-
-declare module "path-is-absolute" {
-    namespace pathIsAbsolute {}
-    function pathIsAbsolute(path: string): boolean;
-    export = pathIsAbsolute;
-}