+++ src/ruleLoader.ts
@@ -52,8 +52,13 @@ export function loadRules(ruleConfiguration: {[name: string]: any},
     }
 
     if (notFoundRules.length > 0) {
-        const errorMessage = `Could not find the following rules specified in the configuration:\n${notFoundRules.join("\n")}`;
-        throw new Error(errorMessage);
+        const ERROR_MESSAGE = `
+            Could not find implementations for the following rules specified in the configuration:
+            ${notFoundRules.join("\n")}
+            Try upgrading TSLint and/or ensuring that you have all necessary custom rules installed.
+            If TSLint was recently upgraded, you may have old rules configured which need to be cleaned up.
+        `;
+        throw new Error(ERROR_MESSAGE);
     } else {
         return rules;
     }