+++ src/rules/banRule.ts
@@ -19,6 +19,27 @@ import * as ts from "typescript";
 import * as Lint from "../lint";
 
 export class Rule extends Lint.Rules.AbstractRule {
+    /* tslint:disable:object-literal-sort-keys */
+    public static metadata: Lint.IRuleMetadata = {
+        ruleName: "ban",
+        description: "Bans the use of specific functions.",
+        descriptionDetails: "At this time, there is no way to disable global methods with this rule.",
+        optionsDescription: "A list of `['object', 'method']` pairs which ban `object.method()`.",
+        options: {
+            type: "list",
+            listType: {
+                type: "array",
+                arrayMembers: [
+                    { type: "string" },
+                    { type: "string" },
+                ],
+            },
+        },
+        optionExamples: [`[true, ["console", "log"], ["someObject", "someFunction"]]`],
+        type: "functionality",
+    };
+    /* tslint:enable:object-literal-sort-keys */
+
     public static FAILURE_STRING_PART = "function invocation disallowed: ";
 
     public apply(sourceFile: ts.SourceFile): Lint.RuleFailure[] {
@@ -44,19 +65,27 @@ export class BanFunctionWalker extends Lint.RuleWalker {
                 && expression.getChildCount() >= 3) {
 
             const firstToken = expression.getFirstToken();
-            const secondToken = expression.getChildAt(1);
-            const thirdToken = expression.getChildAt(2);
+            const firstChild = expression.getChildAt(0);
+            const secondChild = expression.getChildAt(1);
+            const thirdChild = expression.getChildAt(2);
 
-            const firstText = firstToken.getText();
-            const thirdText = thirdToken.getFullText();
+            const rightSideExpression = thirdChild.getFullText();
 
-            if (secondToken.kind === ts.SyntaxKind.DotToken) {
+            let leftSideExpression: string;
+
+            if (firstChild.getChildCount() > 0) {
+                leftSideExpression = firstChild.getLastToken().getText();
+            } else {
+                leftSideExpression = firstToken.getText();
+            }
+
+            if (secondChild.kind === ts.SyntaxKind.DotToken) {
                 for (const bannedFunction of this.bannedFunctions) {
-                    if (firstText === bannedFunction[0] && thirdText === bannedFunction[1]) {
+                    if (leftSideExpression === bannedFunction[0] && rightSideExpression === bannedFunction[1]) {
                         const failure = this.createFailure(
                             expression.getStart(),
                             expression.getWidth(),
-                            `${Rule.FAILURE_STRING_PART}${firstText}.${thirdText}`
+                            `${Rule.FAILURE_STRING_PART}${leftSideExpression}.${rightSideExpression}`
                         );
                         this.addFailure(failure);
                     }