+++ src/rules/useIsnanRule.ts
@@ -22,6 +22,20 @@ import * as Lint from "../lint";
  * Implementation of the use-isnan rule.
  */
 export class Rule extends Lint.Rules.AbstractRule {
+    /* tslint:disable:object-literal-sort-keys */
+    public static metadata: Lint.IRuleMetadata = {
+        ruleName: "use-isnan",
+        description: "Enforces use of the `isNaN()` function to check for NaN references instead of a comparison to the `NaN` constant.",
+        rationale: Lint.Utils.dedent`
+            Since \`NaN !== NaN\`, comparisons with regular operators will produce unexpected results.
+            So, instead of \`if (myVar === NaN)\`, do \`if (isNaN(myVar))\`.`,
+        optionsDescription: "Not configurable.",
+        options: null,
+        optionExamples: ["true"],
+        type: "functionality",
+    };
+    /* tslint:enable:object-literal-sort-keys */
+
     public static FAILURE_STRING = "Found an invalid comparison for NaN: ";
 
     public apply(sourceFile: ts.SourceFile): Lint.RuleFailure[] {
@@ -30,9 +44,9 @@ export class Rule extends Lint.Rules.AbstractRule {
 }
 
 class UseIsnanRuleWalker extends Lint.RuleWalker {
-
     protected visitBinaryExpression(node: ts.BinaryExpression): void {
-        if (this.isExpressionNaN(node.left) || this.isExpressionNaN(node.right)) {
+        if ((this.isExpressionNaN(node.left) || this.isExpressionNaN(node.right))
+                && node.operatorToken.kind !== ts.SyntaxKind.EqualsToken) {
             this.addFailure(this.createFailure(node.getStart(), node.getWidth(), Rule.FAILURE_STRING + node.getText()));
         }
         super.visitBinaryExpression(node);