+++ test/assert.ts
@@ -14,13 +14,13 @@
  * limitations under the License.
  */
 
+/* tslint:disable */
 declare namespace NodeJS {
     interface Global {
         assert: Chai.Assert;
     }
 }
 
-/* tslint:disable:no-var-keyword */
 var assert: Chai.Assert = require("chai").assert;
 global.assert = assert;
-/* tslint:enable:no-var-keyword */
+/* tslint:enable */