+++ test/check-bin.sh
@@ -20,7 +20,7 @@ expectOut () {
   msg=$3
 
   nodeV=`node -v`
-  
+
   # if Node 0.10.*, node will sometimes exit with status 8 when an error is thrown
   if [[ $expect != $actual || $nodeV == v0.10.* && $expect == 1 && $actual == 8 ]] ; then
     echo "$msg: expected $expect got $actual"
@@ -70,6 +70,10 @@ expectOut $? 0 "tslint with with -r pointing to custom rules did not find lint f
 ./bin/tslint -c test/config/tslint-almost-empty.json src/tslint.ts
 expectOut $? 0 "-c relative path without ./ did not work"
 
+# make sure calling tslint with a config file which extends a package relative to the config file works
+./bin/tslint -c test/config/tslint-extends-package-no-mod.json src/tslint.ts
+expectOut $? 0 "tslint (with config file extending relative package) did not work"
+
 # make sure tslint --init generates a file
 cd ./bin
 if [ -f tslint.json ]; then
@@ -96,6 +100,13 @@ expectOut $? 0 "tslint --test did not exit correctly for a passing test"
 
 ./bin/tslint --test test/files/incorrect-rule-test
 expectOut $? 1 "tslint --test did not exit correctly for a failing test"
+
+# ensure --test command works correctly with custom rules
+./bin/tslint --test test/files/custom-rule-rule-test
+expectOut $? 0 "tslint --test did not exit correctly for a passing test with custom rules"
+
+./bin/tslint -r test/files/custom-rules-2 --test test/files/custom-rule-cli-rule-test
+expectOut $? 0 "tslint --test did not exit correctly for a passing test with custom rules from the CLI"
 
 if [ $num_failures != 0 ]; then
   echo "Failed $num_failures tests"